+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Change in import project, switch button attachinary_file_field to file_field
 - Change the verification to refuse if const files is null, otherwise return the value files
 - Request signature for file uploads once and all stories get this from a global place
+- the function handleBackLoggedRelease call only when change estimate or position of some story
 
 ## [1.17.0] 2018-01-23
 ### Fixed